completes Token static methods and fixes users controller logic

+++ app/models/token.rb
@@ -15,17 +15,23 @@ class Token < ActiveRecord::Base
     return new_token
   end
 
-  def self.consume (nonce)
+  def self.consume (nonce, id)
     # Try to find a matching token object by the nonce given
     token = Token.find_by(nonce: nonce)
     # If token exists
     if token != nil
-      # Find the associated user by their id in the token object
-      user = User.find_by(id: token[:user_id])
-      # Destroy the token
-      token.destroy
-      # Return the user object
-      return user
+      # and if it belongs to the current user
+      if token[:user_id] == id
+        # Find the associated user by their id in the token object
+        user = User.find_by(id: token[:user_id])
+        # Destroy the token
+        token.destroy
+        # Return the user object
+        return user
+      else
+        # If that token does not belong to the current user
+        return nil
+      end
     else
       # If the token does not exist...
       return nil
